requirements.txt
anthropic>=0.39.0
pdfplumber>=0.11.0
pandas>=2.2.0
python-docx>=1.1.0
pytesseract>=0.3.10
pillow>=10.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
tqdm>=4.66.0
